requirements.txt
djangorestframework
djangorestframework-simplejwt
python-decouple
django-cors-headers